mental mass 2/backend/requirements.txt
Flask
flask-cors
flask-jwt-extended
opencv-python
deepface
transformers
torch
scikit-learn
reportlab
werkzeug
Pillow
openai
fer